+++ MovieCatalog1/app/src/main/java/com/hafidmust/moviecatalog1/data/tv/TvEntity.kt
@@ -35,7 +35,7 @@ data class TvEntity(
 	val popularity: Double? = null,
 
 	@field:SerializedName("vote_average")
-	val voteAverage: Double? = null,
+	val voteAverage: Int? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
